+++ out/extension.js
@@ -6,6 +6,7 @@ const vscode = require("vscode");
 const mapEditor_1 = require("./editors/mapEditor");
 const characterEditor_1 = require("./editors/characterEditor");
 const itemEditor_1 = require("./editors/itemEditor");
+const spellEditor_1 = require("./editors/spellEditor");
 const notesEditor_1 = require("./editors/notesEditor");
 const statBlockEditor_1 = require("./editors/statBlockEditor");
 const hoverProvider_1 = require("./providers/hoverProvider");
@@ -24,6 +25,7 @@ function activate(context) {
     context.subscriptions.push(mapEditor_1.MapEditorProvider.register(context));
     context.subscriptions.push(characterEditor_1.CharacterSheetProvider.register(context));
     context.subscriptions.push(itemEditor_1.ItemEditorProvider.register(context));
+    context.subscriptions.push(spellEditor_1.SpellEditorProvider.register(context));
     context.subscriptions.push(notesEditor_1.NotesEditorProvider.register(context));
     context.subscriptions.push(statBlockEditor_1.StatBlockEditorProvider.register(context));
     // Register Hover Provider